Mock all versions of tomcat for testing
