+++ conftest.py
@@ -8,7 +8,6 @@ import os
 import pytest
 
 import tomcatmanager as tm
-from tests.mock_server_10_0 import start_mock_server_10_0
 
 ###
 #
@@ -33,28 +32,35 @@ class TomcatServer:
 ###
 def pytest_addoption(parser):
     parser.addoption(
+        "--mocktomcat",
+        action="store",
+        default=tm.TomcatMajorMinor.highest_supported().value,
+        choices=list(map(lambda v: v.value, tm.TomcatMajorMinor.supported())),
+        help="test against a specific mock version of Tomcat",
+    )
+    parser.addoption(
         "--url",
         action="store",
         default=None,
-        help="url: url of tomcat manager to test against instead of mock",
+        help="url of tomcat manager to test against instead of mock server",
     )
     parser.addoption(
-        "--user", action="store", default=None, help="user: use to authenticate"
+        "--user", action="store", default=None, help="use to authenticate at url"
     )
     parser.addoption(
-        "--password", action="store", default=None, help="password: use to authenticate"
+        "--password", action="store", default=None, help="use to authenticate at url"
     )
     parser.addoption(
         "--warfile",
         action="store",
         default=None,
-        help="warfile: path to deployable war file on the tomcat server",
+        help="path to deployable war file on the tomcat server",
     )
     parser.addoption(
         "--contextfile",
         action="store",
         default=None,
-        help="contextfile: path to context.xml file on the tomcat server",
+        help="path to context.xml file on the tomcat server",
     )
 
 
@@ -117,7 +123,27 @@ def tomcat_manager_server(request):
         return tms
     else:
         # go start up a fake server
-        return start_mock_server_10_0(tms)
+        mockver = request.config.getoption("--mocktomcat")
+        if mockver == tm.TomcatMajorMinor.V10_0.value:
+            from tests.mock_server_10_0 import start_mock_server_10_0
+
+            return start_mock_server_10_0(tms)
+        elif mockver == tm.TomcatMajorMinor.V9_0.value:
+            from tests.mock_server_9_0 import start_mock_server_9_0
+
+            return start_mock_server_9_0(tms)
+        elif mockver == tm.TomcatMajorMinor.V8_5.value:
+            from tests.mock_server_8_5 import start_mock_server_8_5
+
+            return start_mock_server_8_5(tms)
+        elif mockver == tm.TomcatMajorMinor.V8_0.value:
+            from tests.mock_server_8_0 import start_mock_server_8_0
+
+            return start_mock_server_8_0(tms)
+        elif mockver == tm.TomcatMajorMinor.V7_0.value:
+            from tests.mock_server_7_0 import start_mock_server_7_0
+
+            return start_mock_server_7_0(tms)
 
 
 @pytest.fixture
