+++ tests/mock_server_8_5.py
@@ -24,7 +24,7 @@
 # pylint: disable=too-many-lines, too-many-public-methods
 
 """
-Mock up a Tomcat Manager application that behaves like tomcat version 10.0.x
+Mock up a Tomcat Manager application that behaves like tomcat version 8.5.x
 """
 
 from http.server import HTTPServer
@@ -34,14 +34,14 @@ import threading
 from tests.mock_server_ssl import MockRequestHandlerSSL
 
 
-class MockRequestHandler100(MockRequestHandlerSSL):
-    """Handle HTTP Requests like Tomcat Manager 10.0.x"""
+class MockRequestHandler85(MockRequestHandlerSSL):
+    """Handle HTTP Requests like Tomcat Manager 8.5.x"""
 
     def get_server_info(self):
         """Send the server information."""
         self.send_text(
             """OK - Server info
-Tomcat Version: [Apache Tomcat/10.0.4]
+Tomcat Version: [Apache Tomcat/8.5.65]
 OS Name: [Linux]
 OS Version: [5.4.0-67-generic]
 OS Architecture: [amd64]
@@ -54,7 +54,7 @@ JVM Vendor: [Private Build]"""
 #
 #
 ###
-def start_mock_server_10_0(tms):
+def start_mock_server_8_5(tms):
     """Start a mock Tomcat Manager application
 
     :return: a tuple: (url, user, password) where the server is accessible
@@ -67,13 +67,13 @@ def start_mock_server_10_0(tms):
     sock.close()
 
     tms.url = "http://localhost:{}/manager".format(port)
-    tms.user = MockRequestHandler100.USER
-    tms.password = MockRequestHandler100.PASSWORD
+    tms.user = MockRequestHandler85.USER
+    tms.password = MockRequestHandler85.PASSWORD
     tms.warfile = "/path/to/server.war"
     tms.contextfile = "path/to/context.xml"
     tms.connect_command = "connect {} {} {}".format(tms.url, tms.user, tms.password)
 
-    mock_server = HTTPServer(("localhost", port), MockRequestHandler100)
+    mock_server = HTTPServer(("localhost", port), MockRequestHandler85)
     mock_server_thread = threading.Thread(target=mock_server.serve_forever)
     mock_server_thread.daemon = True
     mock_server_thread.start()
