+++ tests/test_manager_ssl.py
@@ -28,10 +28,15 @@ import unittest.mock as mock
 import pytest
 import tomcatmanager as tm
 
+
 def test_ssl_connector_ciphers(tomcat, assert_tomcatresponse):
-    r = tomcat.ssl_connector_ciphers()
-    assert_tomcatresponse.info(r)
-    assert r.result == r.ssl_connector_ciphers
+    if tomcat.implements(tomcat.ssl_connector_ciphers):
+        r = tomcat.ssl_connector_ciphers()
+        assert_tomcatresponse.info(r)
+        assert r.result == r.ssl_connector_ciphers
+    else:
+        with pytest.raises(tm.TomcatNotImplementedError):
+            r = tomcat.ssl_connector_ciphers()
 
 
 def test_ssl_connector_certs(tomcat, assert_tomcatresponse):
@@ -61,13 +66,16 @@ def test_ssl_reload_success(tomcat, mocker, assert_tomcatresponse):
         mock_result = mocker.patch(
             "requests.Response.text", create=True, new_callable=mock.PropertyMock
         )
-        mock_result.return_value = "OK - Reloaded TLS configuration for [www.example.com]"
+        mock_result.return_value = (
+            "OK - Reloaded TLS configuration for [www.example.com]"
+        )
         r = tomcat.ssl_reload("www.example.com")
         assert_tomcatresponse.success(r)
         assert r.status_message == "Reloaded TLS configuration for [www.example.com]"
     else:
         with pytest.raises(tm.TomcatNotImplementedError):
             r = tomcat.ssl_reload("www.example.com")
+
 
 def test_ssl_reload_fail(tomcat, mocker, assert_tomcatresponse):
     # the command on the tomcat manager web app fails if SSL is not configured
